Implement swizzled access.
This also adds a new CONSTRUCT_PRIMITIVE instruction which does not require
multiple array access instructions to build a primitive, and a new SHUFFLE
instruction similar to the one in SPIR-V.

Additionally, the visitor has been refactor: It now calls OnEnter/OnLeave while
visiting, which allows tracking state in the context. This is now used to
handle all assignments -- instead of calling `SetStore` after the fact, the
stores are now emitted while handling the instruction.