Replace uses as well.
Previously, we replaced instructions, but other instructions referencing them
were not notified as we reference instructions directly, and not just their
reference. The replaced instruction will be thus kept alive by references from
other instructions.

To fix this, we add a new method ReplaceUses to Function, BasicBlock, and
Instruction, which replaces uses with a new value. Additionally, we also add
Uses() which returns the used values, which allows us to build a mapping of
value use to instruction -- making replacements much faster (as instead of
iterating over all instructions, we just iterate over the instructions that
actually use a value, which is generally much shorter.)