Prepare everything for module-level importing.
We introduce a new import "module"; statement, which imports an already compiled
module. This makes all types and functions from that module available in the
current module; but does not import globals. This also adds a linker which can
combine multiple modules together.

Additionally, a bug related to overloaded functions has been resolved: If the
parameter list was longer than the argument list, only the arguments were
compared to the parameter list instead of returning "no match".